Add support for fzf-lua
For the colours to apply to the fzf window, you need to pass
fzf_colors = true in the setup function for fzf-lua.

The colour of the text when it is being highlighted is completely
invisible in the light theme, so this fixes this by setting a specific
colour for the text. The colour of which follows the behaviour of
fzf lightening the text colour when it is highlighted.

Also reformatted the README.md file using prettier.